start moving save-related code out of MVStore into FileStore
