+++ h2/src/main/org/h2/mvstore/MVStoreTool.java
@@ -110,7 +110,7 @@ public class MVStoreTool {
         }
         long size = FileUtils.size(fileName);
         pw.printf("File %s, %d bytes, %d MB\n", fileName, size, size / 1024 / 1024);
-        int blockSize = MVStore.BLOCK_SIZE;
+        int blockSize = FileStore.BLOCK_SIZE;
         TreeMap<Integer, Long> mapSizesTotal =
                 new TreeMap<>();
         long pageSizeTotal = 0;
@@ -158,7 +158,7 @@ public class MVStoreTool {
                     pos += blockSize;
                     continue;
                 }
-                int length = c.len * MVStore.BLOCK_SIZE;
+                int length = c.len * FileStore.BLOCK_SIZE;
                 pw.printf("%n%0" + len + "x chunkHeader %s%n",
                         pos, c.toString());
                 ByteBuffer chunk = ByteBuffer.allocate(length);
@@ -363,7 +363,7 @@ public class MVStoreTool {
                 if (k.startsWith(DataUtils.META_CHUNK)) {
                     Chunk c = Chunk.fromString(e.getValue());
                     chunks.put(c.id, c);
-                    chunkLength += c.len * MVStore.BLOCK_SIZE;
+                    chunkLength += c.len * FileStore.BLOCK_SIZE;
                     maxLength += c.maxLen;
                     maxLengthLive += c.maxLenLive;
                     if (c.maxLenLive > 0) {
@@ -622,7 +622,7 @@ public class MVStoreTool {
         }
         FileChannel file = null;
         FileChannel target = null;
-        int blockSize = MVStore.BLOCK_SIZE;
+        int blockSize = FileStore.BLOCK_SIZE;
         try {
             file = FilePath.get(fileName).open("r");
             FilePath.get(fileName + ".temp").delete();
@@ -657,7 +657,7 @@ public class MVStoreTool {
                     pos += blockSize;
                     continue;
                 }
-                int length = c.len * MVStore.BLOCK_SIZE;
+                int length = c.len * FileStore.BLOCK_SIZE;
                 ByteBuffer chunk = ByteBuffer.allocate(length);
                 DataUtils.readFully(file, pos, chunk);
                 if (c.version > targetVersion) {
@@ -673,9 +673,9 @@ public class MVStoreTool {
                 }
                 pos += length;
             }
-            int length = newestChunk.len * MVStore.BLOCK_SIZE;
+            int length = newestChunk.len * FileStore.BLOCK_SIZE;
             ByteBuffer chunk = ByteBuffer.allocate(length);
-            DataUtils.readFully(file, newestChunk.block * MVStore.BLOCK_SIZE, chunk);
+            DataUtils.readFully(file, newestChunk.block * FileStore.BLOCK_SIZE, chunk);
             chunk.rewind();
             target.write(chunk, fileSize);
         } catch (IOException e) {
