+++ h2/src/test/org/h2/test/store/TestMVStore.java
@@ -14,6 +14,7 @@ import java.util.Map.Entry;
 import java.util.NoSuchElementException;
 import java.util.Random;
 import java.util.TreeMap;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
@@ -153,9 +154,9 @@ public class TestMVStore extends TestBase {
         FileStore fileStore = new OffHeapStore() {
 
             @Override
-            public void open(String fileName, boolean readOnly, char[] encryptionKey) {
+            public void open(String fileName, boolean readOnly, char[] encryptionKey, ConcurrentHashMap<Integer, Chunk> chunks) {
                 openClose.incrementAndGet();
-                super.open(fileName, readOnly, encryptionKey);
+                super.open(fileName, readOnly, encryptionKey, chunks);
             }
 
             @Override
